Removed unused code, updated sequence-files.ts to use requests file for get/post and updated to use typing
