+++ src/main/webapp/resources/js/components/samples/components/fastqc/fastQCSlice.ts
@@ -1,35 +1,6 @@
 import { createAction, createSlice } from "@reduxjs/toolkit";
 import { SequencingFile } from "../../../../apis/samples/samples";
-
-export interface FastQC {
-  additionalProperties: Record<string, unknown>[];
-  analysisType: Record<string, string>;
-  createdDate: Date;
-  description: string;
-  encoding: string;
-  executionManagerAnalysisId: string;
-  fastqcVersion: string;
-  fileType: string;
-  filteredSequences: number;
-  gcContent: number;
-  identifier: number;
-  label: string;
-  links: any[];
-  maxLength: number;
-  minLength: number;
-  overrepresentedSequences: OverrepresentedSequences[];
-  totalBases: number;
-  totalSequences: number;
-}
-
-export interface OverrepresentedSequences {
-  sequences: string;
-  overrepresentedSequenceCount: number;
-  percentage: number;
-  possibleSource: string;
-  createdDate: Date;
-  identifier: number;
-}
+import { FastQC } from "../../../../apis/files/sequence-files";
 
 /**
  * Action to set the fast qc details
